Move common loader functionality to AbstractTileLoader
